add comments to types
