+++ annotations.go
@@ -1,5 +1,6 @@
 package fir
 
+// CreateForm can be used to annotate an entgo schema field as input in a create form.
 type CreateForm struct {
 	Fields []string
 }
@@ -8,6 +9,7 @@ func (CreateForm) Name() string {
 	return "CreateForm"
 }
 
+// UpdateForm can be used to annotate an entgo schema field as input in an update form.
 type UpdateForm struct {
 	Fields []string
 }
@@ -16,6 +18,7 @@ func (UpdateForm) Name() string {
 	return "UpdateForm"
 }
 
+// ListItem can be used to annotate an entgo schema field as a member of a list item.
 type ListItem struct {
 	Fields []string
 }
