+++ docs/content/en/docs/api/go.md
@@ -16,17 +16,48 @@ toc: true
 
 
 ## View
-
-### View interface
-### DefaultView
-### Page
-### Patch
-#### Patchset
-
+### ID
+### Content
+### Layout
+### LayoutContentName
+### Partials
+### Extensions
+### FuncMap
+### OnGet
+### OnPost
+### OnEvent
+### Publisher
+## DefaultView
+## DefaultErrorView
+## Page
+### Data
+### Code
+### Message
+### Error
+## PageContext
+## Data
+## Patch
+### Op
+### GetSelector
+### GetTemplate
+## List of patch operations
+### After
+### Append
+### Before
+### Morph
+### Navigate
+### Prepend
+### Reload
+### Remove
+### ResetForm
+### Store
+## Patchset
+## Event
+### ID
+### Params
+### DecodeParams
+### DecodeFormParams
+### RequestContext
 ## Controller
+## Controller Options
 
-### Controller interface
-
-### DefaultController
-
-### Options
