+++ errors.go
@@ -15,6 +15,7 @@ func getUserMessage(status int, userMessage []string) string {
 	return msg
 }
 
+// ErrInternalServer returns a Page with an internal server error.
 func ErrInternalServer(err error, userMessage ...string) Page {
 	return Page{
 		Code:    http.StatusInternalServerError,
@@ -23,6 +24,7 @@ func ErrInternalServer(err error, userMessage ...string) Page {
 	}
 }
 
+// ErrBadRequest returns a Page with a bad request error.
 func ErrBadRequest(err error, userMessage ...string) Page {
 	return Page{
 		Code:    http.StatusBadRequest,
@@ -31,10 +33,20 @@ func ErrBadRequest(err error, userMessage ...string) Page {
 	}
 }
 
+// ErrNotFound returns a Page with a not found error.
 func ErrNotFound(err error, userMessage ...string) Page {
 	return Page{
 		Code:    http.StatusNotFound,
 		Message: getUserMessage(http.StatusNotFound, userMessage),
+		Error:   err,
+	}
+}
+
+// ErrUnauthorized returns a Page with an unauthorized error.
+func ErrUnauthorized(err error, userMessage ...string) Page {
+	return Page{
+		Code:    http.StatusUnauthorized,
+		Message: getUserMessage(http.StatusUnauthorized, userMessage),
 		Error:   err,
 	}
 }
@@ -48,6 +60,7 @@ func morphError(err string) Patch {
 	}
 }
 
+// PatchError returns a patchset that sets an error for selector #fir-error.
 func PatchError(err error, userMessage ...string) Patchset {
 	msg := "internal error"
 	if err != nil && len(userMessage) == 0 {
@@ -62,6 +75,7 @@ func PatchError(err error, userMessage ...string) Patchset {
 	return Patchset{morphError(msg)}
 }
 
+// PageError returns a Page with an error.
 func PageError(err error, userMessage ...string) Page {
 	msg := "internal error"
 	if err != nil && len(userMessage) == 0 {
@@ -84,6 +98,7 @@ func PageError(err error, userMessage ...string) Page {
 	}
 }
 
+// UnsetFormError returns a patchset that unsets the error for a form.
 func UnsetFormErrors(fields ...string) Patchset {
 	var patchset Patchset
