+++ helpers.go
@@ -8,6 +8,7 @@ import (
 
 var decoder = schema.NewDecoder()
 
+// DecodeForm decodes the form values from the request into the destination struct.
 func DecodeForm(dst interface{}, r *http.Request) error {
 	err := r.ParseForm()
 	if err != nil {
@@ -22,6 +23,7 @@ func DecodeForm(dst interface{}, r *http.Request) error {
 	return nil
 }
 
+// DecodeURLValues decodes the url values from the request into the destination struct.
 func DecodeURLValues(dst interface{}, r *http.Request) error {
 	err := decoder.Decode(dst, r.URL.Query())
 	if err != nil {
