+++ page.go
@@ -8,20 +8,20 @@ type Page struct {
 	Error   error  `json:"-"`
 }
 
-// AppContext is a struct that holds the data for the app context
-type AppContext struct {
+// PageContext is a struct that holds the data for the app context
+type PageContext struct {
 	Name    string
 	URLPath string
 }
 
-func (a *AppContext) ActiveRoute(path, class string) string {
+func (a *PageContext) ActiveRoute(path, class string) string {
 	if a.URLPath == path {
 		return class
 	}
 	return ""
 }
 
-func (a *AppContext) NotActiveRoute(path, class string) string {
+func (a *PageContext) NotActiveRoute(path, class string) string {
 	if a.URLPath != path {
 		return class
 	}
