+++ watch.go
@@ -13,6 +13,7 @@ import (
 	"golang.org/x/exp/slices"
 )
 
+// DefaultWatchExtensions is an array of default extensions to watch for changes.
 var DefaultWatchExtensions = []string{".gohtml", ".gotmpl", ".html", ".tmpl"}
 
 const devReloadChannel = "dev_reload"
